fix: add typehints for input csv dataframes
When reading in the programs CSV files, the columns containing sparse
numerical values where being read in as floats as pandas does not
normally allow for NA values for int types. This then required a
later step to resave the file from Excel to strip the float types.

Since the sparse columns are more like enumerated types, with more
symbolic than numeric meaning, we can instead read in the types as
string values instead of numeric values. This prevents coercion to
floating types and removes the need to resave the file in Excel.

Additionally, this fixes an issue where leading zeros for phone
extensions where being cut off.