fix: first single quote escape for fixtured data
There was an error in how the first single quote for data being pulled
into the data tables was being handled. It was converting the first
single quote into a double quotation mark, instead of the normal
process of converting the single quotes into two single quotes for
sql string escaping.