chore: reduce logging during feature tests
When running feature tests in CircleCI, the migration scripts running
on a clean database result in around 300MBs of logged output because
db-migrate logs all the data from the sql files. This makes it
impractical to review failed test output on CircleCI failures.

This redirects the output of starting up the local prod server during
feature tests to the null device to swallow the migration logs.