fix: remove line breaks that appear to be in middle of sentences
In the source data, there are description records that seem to be
copy and paste entries from a PDF or other program that created
line breaks in the middle of sentences (likely the end of the line
when displayed in the original program). Examples are for trainings
1663 and 9381.

This change removes line breaks if they are the only thing between
two alphabetical characters. While there are some false matches, the
overall result does look better.