Use conditional logging based on NODE_ENV
Updated the logging mechanism to conditionally utilize cloudwatchLogger and sentryLogger only when NODE_ENV is set to 'awsprod', 'awstest', or 'awsdev'. In other environments, the system defaults to the original console logging. This change ensures that enhanced logging services are engaged only in specified deployment environments, avoiding potential issues in development or other non-AWS environments.