Exclude in-demand SOCs from local exceptions and support 2010 SOC codes
This commit updates the `getLocalExceptionsBySoc` method in the `PostgresDataClient` class to exclude Standard Occupational Classification (SOC) codes that are marked as in-demand in the `indemandsocs` table. It enhances the database query to filter out occupations listed in this table, ensuring that the returned local exceptions only include occupations not currently in high demand.

Additionally, the method has been expanded to support both 2010 and 2018 SOC codes by incorporating a join with a crosswalk table (`soc2010to2018crosswalk`). This allows the method to recognize and properly handle requests for local exceptions based on either version of SOC codes, increasing the utility and flexibility of the query for users working with different datasets.

Modifications include:
- Using a `LEFT JOIN` with the `indemandsocs` table to exclude in-demand SOCs.
- Joining with the `soc2010to2018crosswalk` table to map 2018 SOC codes to their 2010 counterparts and vice versa.
- Filtering out occupations that are in demand for both 2010 and 2018 SOC codes by checking for null values in the `indemandsocs` table and its joined counterpart for 2010 codes.

These changes ensure that the method more accurately reflects the current labor market's demands while providing compatibility with datasets that use either version of the SOC system.