fix(deps): Resolve react-scripts start error
Removed the `webpack-dev-server` override from `package.json`.

The override forced an installation of `webpack-dev-server` v5.x, which is incompatible with `react-scripts@5.0.1`. This version of `react-scripts` uses the `onAfterSetupMiddleware` option, which was removed in `webpack-dev-server` v5, causing the development server to crash on startup.

Removing the override allows npm to install a compatible version of `webpack-dev-server` (v4), fixing the issue.